add get command handling
